Update Day08.kt and Day08.txt
This commit adds the implementation for the Day08 task. It includes functions for parsing the input, calculating the number of steps based on the given route and map, and solving both parts of the task. Some helper functions for finding the number of steps between start and end nodes and calculating the least common multiple (LCM) are also included.

In addition, some test cases are provided to validate the implementation's correctness. The main function of the program reads the input file and calls the part1 and part2 functions to solve the task. The results are printed to the console.

The new file Day08.txt contains a long string of directions and map data for various nodes. The format of the map data is a tuple with two strings representing the left and right connections for each node.

The implementation of the LCM function uses the Euclidean algorithm to calculate the greatest common divisor (GCD) of two numbers, which is then used to calculate the LCM.